pyproject.toml
[project]
name = "drivedial"
version = "1.0.0"
description = "AI Voice Sales Agent - Django + Twilio + OpenAI Realtime"
requires-python = ">=3.9"

dependencies = [
    "aiohttp>=3.13.0",
    "channels>=4.0.0",
    "channels-redis>=4.0.0",
    "daphne>=4.0.0",
    "django>=4.2.0",
    "django-cors-headers>=4.0.0",
    "djangorestframework>=3.14.0",
    "pyngrok>=7.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "twisted[http2,tls]>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
asyncio_mode = "auto"
